Resolution of one potential bug.

+++ src_dualdesc/POLY_SerialDualDesc.cpp
@@ -15,7 +15,8 @@ int main(int argc, char *argv[])
     }
     std::string eFileName=argv[1];
     using T = mpq_class;
-    using Tidx = uint8_t;
+    //    using Tidx = uint8_t;
+    using Tidx = uint16_t;
     using Telt = permutalib::SingleSidedPerm<Tidx>;
     using Tint = mpz_class;
     using Tgroup = permutalib::Group<Telt,Tint>;
